+++ other/inversions.cpp
@@ -1,8 +1,8 @@
 template <typename T>
-int64_t merge_sort(T *v, T *aux, int l, int r) {
+long long merge_sort(T *v, T *aux, int l, int r) {
   if (r <= l + 1) return 0;
   int m = (l + r) / 2;
-  int64_t res = merge_sort(v, aux, l, m) + merge_sort(v, aux, m, r);
+  long long res = merge_sort(v, aux, l, m) + merge_sort(v, aux, m, r);
   int i = l, j = m, k = l;
   while (i < m && j < r) {
     aux[k++] = v[i] <= v[j] ? v[i++] : (res += m - i, v[j++]);
@@ -16,7 +16,7 @@ int64_t merge_sort(T *v, T *aux, int l, int r) {
   return res;
 }
 template <typename T>
-int64_t count_inversions(std::vector<T> v) {
+long long count_inversions(std::vector<T> v) {
   std::vector<T> aux(v.size());
   return merge_sort(v.data(), aux.data(), 0, v.size());
 }
